Version 1.11.10
Massive changes.
Can get data from other data access if it has a dll that have method
which is marked by the attribute in protocol and return a datatable in
it.
Almost every method is parallel now and every method is executed by the
data form ,not the result form any more.
The cost time and method name can be shown at the main form.